recommendation-service/requirements.txt
flask==3.0.0
pymongo==4.5.0
pika==1.3.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1